+++ pom.xml
@@ -10,7 +10,7 @@
     </parent>
     <groupId>com.rebuild</groupId>
     <artifactId>rebuild</artifactId>
-    <version>2.2.0</version>
+    <version>2.3.0-dev</version>
     <name>rebuild</name>
     <description>RB V2 use SpringBoot</description>
     <!-- UNCOMMENT USE TOMCAT -->
