+++ src/main/java/com/rebuild/utils/AppUtils.java
@@ -11,7 +11,7 @@ import cn.devezhao.commons.ThrowableUtils;
 import cn.devezhao.commons.web.ServletUtils;
 import cn.devezhao.commons.web.WebUtils;
 import cn.devezhao.persist4j.engine.ID;
-import com.rebuild.api.user.AuthTokenManager;
+import com.rebuild.api.AuthTokenManager;
 import com.rebuild.core.Application;
 import com.rebuild.core.BootApplication;
 import com.rebuild.core.support.ConfigurationItem;
